backend/requirements.txt
# --- Core Web & Agent Frameworks ---
fastapi==0.111.0
uvicorn[standard]==0.29.0  # [standard] includes performance-enhancing libraries
pydantic>=2.7.1
langchain>=0.2.1
langgraph>=0.0.60
langchain-community>=0.2.1
langchain-core>=0.2.1
langchain-mcp-adapters>=0.2.1

# --- Local LLM & Vector DB Integration ---
ollama>=0.2.0
chromadb>=0.5.0
langchain-ollama>=0.2.1

# --- Security & Credential Management ---
cryptography>=42.0.7 # For the fallback encrypted credential vault (Fernet)

# --- Environment & Configuration ---
python-dotenv>=1.0.1

# --- API Interaction ---
httpx>=0.27.0 # Modern async HTTP client, often used by langchain

